STORAGE: moved toolchain_support.h into platform/retarget.h
